LayoutTests/imported/w3c/web-platform-tests/IndexedDB/idbrequest_error-expected.txt


PASS IDBRequest.error throws if ready state is pending
